+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 
 * Pre-defined paper sizes of the ISO A, B and C series to be more precise
 
+### Fixed
+
+* [HexaPDF::Layout::Box#fit] to use float comparison
+
 
 ## 0.35.0 - 2024-01-06